+++ src/main/java/utilities/AllureReport.java
@@ -45,7 +45,7 @@ public class AllureReport {
             .append("<p><b>Platform Version:</b>".concat(isAndroid() ? "Android " : "IOS"))
             .append(platformVersion)
             .append("</p>");
-        if (!BasePage.isNullOrEmpty(appActivity)) {
+        if (!FrontEndOperation.isNullOrEmpty(appActivity)) {
           description.append("<p><b>App Activity:</b> ").append(appActivity).append("</p>");
         }
       } else {
@@ -64,7 +64,7 @@ public class AllureReport {
           .append(LocalEnviroment.getAppIdentifier())
           .append("</p>");
       String apk = LocalEnviroment.getApk();
-      if (!BasePage.isNullOrEmpty(apk)) {
+      if (!FrontEndOperation.isNullOrEmpty(apk)) {
         description.append("<p><b>Apk:</b> ").append(apk).append("</p>");
       }
     } else {
