--- src/main/resources/texts/texts_en_US.properties
@@ -1,2 +1,0 @@
-greeting=Hello!
-farewell=Goodbye!
