--- src/main/resources/texts/texts_es_ES.properties
@@ -1,2 +1,0 @@
-greeting=¡Hola!
-farewell=¡Adiós!
