--- src/main/resources/texts/texts_fr_FR.properties
@@ -1,2 +1,0 @@
-greeting=Bonjour!
-farewell=Au revoir!
